Modularize Rust code
